Add note clarifying the use of torch-provided luarocks
As described in #62, on OSX is is common that `luarocks` will resolve
to the homebrew-installed luarocks or some other executable than that
provided by torch. This PR proposes adding a note to clarify that the
executable installed by torch should be used in the event that a
dependecy issue is encountered; when I ran into this I thought to
look through closed issues, but that may not be immediately obvious
to all users.